Move plugin spec loaders into plugins service
Delete the stale internal/openapi and internal/graphql packages by moving plugin OpenAPI and GraphQL spec loading under services/plugins. The loaders already produce declarative plugin definitions, so this keeps plugin-owned catalog construction in the plugin service tree.

Interface diff:

- import "github.com/valon-technologies/gestalt/server/internal/openapi"

- openapi.LoadDefinition(ctx, name, specURL, allowedOps)

- import "github.com/valon-technologies/gestalt/server/internal/graphql"

- graphql.LoadDefinition(ctx, name, endpoint, allowedOps, selectionOverrides)

+ import "github.com/valon-technologies/gestalt/server/services/plugins/openapi"

+ openapi.LoadDefinition(ctx, name, specURL, allowedOps)

+ import "github.com/valon-technologies/gestalt/server/services/plugins/graphql"

+ graphql.LoadDefinition(ctx, name, endpoint, allowedOps, selectionOverrides)

Example:

def, err := openapi.LoadDefinition(ctx, "slack", specURL, allowedOps)

def, err := graphql.LoadDefinition(ctx, "linear", endpoint, allowedOps, selections)
