+++ gestaltd/internal/bootstrap/graphql_session_provider_test.go
@@ -14,8 +14,8 @@ import (
 	"github.com/valon-technologies/gestalt/server/core"
 	"github.com/valon-technologies/gestalt/server/core/catalog"
 	"github.com/valon-technologies/gestalt/server/internal/config"
-	"github.com/valon-technologies/gestalt/server/internal/graphql"
 	"github.com/valon-technologies/gestalt/server/services/plugins/declarative"
+	"github.com/valon-technologies/gestalt/server/services/plugins/graphql"
 )
 
 func TestGraphQLSessionCatalogProviderLoadsCatalogOnDemand(t *testing.T) {
