+++ gestaltd/services/plugins/graphql/query.go
