+++ gestaltd/services/plugins/openapi/loader.go
